fixup: fix incorrect parent PID detection bug
When Async ZLE helper code starts a coprocess it starts it in a
subshell. This was at crossroads with assumption that the parent of the
GoPrompt process is the user Shell. To remedy that add
`--pid-parent-skip` option which is set to 1 to skip the extra subshell
and render correct parent process name.

NOTE: knowing the parent process is pretty handy to know if your
session is nested.